Changes index type for mesh back to uint32.

+++ skybox.go
@@ -60,7 +60,7 @@ func NewSkybox(cubeMap *Texture, shader *Program, far float64) *Skybox {
 		1, -1, -1,
 		1, 1, -1,
 	}
-	indicies := []uint16 {
+	indicies := []uint32 {
 		0, 1, 2, 0, 2, 3,
 		4, 5, 6, 4, 6, 7,
 		8, 9, 10, 8, 10, 11,
